Refactor dbconn
One reason pygresql was previously modified was that it did not handle
closing a connection very gracefully. In the process of updating
pygresql, we've wrapped the connection it provides with a
ClosingConnection function, which should handle gracefully closing the
connection when the "with dbconn.connect as conn" syntax is used.

This did, however, illustrate issues where a cursor might have been
created as the result of a dbconn.execSQL() call, which seems to hold
the connection open if not specifically closed.

It is therefore necessary to remove the ability to get a cursor from
dbconn.execSQL(). To highlight this difference, and to ensure that
future calls to this library is easy to use, I've cleaned up and
clarified the dbconn execution code, to include the following features.

- dbconn.execSQL() closes the cursor as part of the function. It returns
  no rows
- functions dbconn.query() is added, which behaves like dbconn.execSQL()
  except that it now returns a cursor
- function dbconn.execQueryforSingleton() is renamed
  dconn.querySingleton()
- function dbconn.execQueryforSingletonRow() is renamed
  dconn.queryRow()

Changes to the original commit:
We needed almost no changes from original commit, but to preserve authorship.
So the only thing that is left it is openning and closing connection in
`analyzedb_mgmt_utils.py` in "rows are inserted into table".
Also connection was removed from context in `create_database_if_not_exists`.

Authored-by: Tyler Ramer <tramer@pivotal.io>
(cherry picked from commit 330db2305749952e534178f37a0dfb35758e17d7)