Completely rewrite the password reuse policies feature to be available in all database. This implementation use a dedicated shared memory storage to share the paswword history between all database.
The module must be loaded by adding credcheck to shared_preload_libraries
in postgresql.conf, because it requires additional shared memory.
This means that a server restart is needed to add or remove the module.

When credcheck is active, it stores password history across all databases
of the server. To access and manipulate this history, the module provides
a view pg_password_history, and the utility functions pg_password_history_reset()
and pg_password_history_timestamp(). These are not available globally but can be
enabled for a specific database with CREATE EXTENSION credcheck.

The password history is stored in share memory and written to disk in
file $PGDATA/global/pg_password_history to be loaded at startup. The
share memory history size is set to 65535 records by default and can
be adjusted using the credcheck.history_max_size configuration
directive. Change of this GUC require a PostgreSQL restart.