Fix parallelization of limit with master-only tables and functions
When subquery called from the query on distributed table, uses limit
clause which references outer query column, previously it was
incorrectly assumed that inner query can be executed on any segment
even if it uses master-only table or function which collects data from
several segments.

This commit changes condition on which plan mutation is skipped, and
allows to add motions to the results of  FunctionScan if locus of
function is not same as locus of outer query and to the table scans
of master-only table.